+++ src/main/java/waktfolio/domain/repository/like/MemberLikeCustomRepository.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface MemberLikeCustomRepository {
-    BigDecimal countAddCountByContentId(UUID contentId);
+    Long countAddCountByContentId(UUID contentId);
     Long countAddCountByMemberId(UUID memberId);
 }
